% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofthrones.R
\docType{data}
\name{gameofthrones}
\alias{gameofthrones}
\title{Game of Thrones episode data}
\format{A \code{data.frame} containing numerical and factor data.}
\source{
\url{https://trakt.tv} via the \code{tRakt} package.
}
\usage{
gameofthrones
}
\description{
Episode data for Game of Thrones retrieved from \url{https://trakt.tv}
}
\note{
Last updated on 2015-12-11
}
\keyword{dataset}

