% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penis.R
\docType{data}
\name{penis}
\alias{penis}
\title{Word Penis Size}
\format{A \code{data.frame} containing numerical and factor data.}
\source{
The horribly coded but very informative \url{http://www.everyoneweb.com/worldpenissize}
}
\usage{
penis
}
\description{
Penis data across multiple regions.
}
\keyword{dataset}

