% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularmovies.R
\docType{data}
\name{popularmovies}
\alias{popularmovies}
\title{Overview of popular movies on \code{trakt.tv}}
\format{A \code{data.frame} containing numerical and factor data.}
\source{
\url{https://trakt.tv} via the \code{tRakt} package.
}
\usage{
popularmovies
}
\description{
A table of 100 movies popular on \url{https://trakt.tv}
}
\note{
Last updated on 2015-12-11
}
\keyword{dataset}

