% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularshows.R
\docType{data}
\name{popularshows}
\alias{popularshows}
\title{Overview of popular shows on \code{trakt.tv}}
\format{
A \code{data.frame} containing numerical and factor data.
}
\source{
\url{https://trakt.tv} via the \code{tRakt} package.
}
\usage{
popularshows
}
\description{
A table of 100 shows popular on \url{https://trakt.tv}
}
\note{
Last updated on 2015-12-11
}
\keyword{dataset}
