% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldrankings.R
\docType{data}
\name{worldrankings}
\alias{worldrankings}
\title{Some world rankings scraped straight from Wikipedia}
\format{A \code{data.frame} containing numerical and character data (country name).}
\source{
\url{https://en.wikipedia.org}.
}
\usage{
worldrankings
}
\description{
A table of various world rankings
\itemize{
\item Literacy rate \url{https://en.wikipedia.org/wiki/List_of_countries_by_literacy_rate}
\item Happiness \url{https://en.wikipedia.org/wiki/World_Happiness_Report}
\item Smartphone penetration \url{https://en.wikipedia.org/wiki/List_of_countries_by_smartphone_penetration}
\item Discrimination \url{https://en.wikipedia.org/wiki/List_of_countries_by_discrimination_and_violence_against_minorities}
\item Homicide \url{https://en.wikipedia.org/wiki/List_of_countries_by_intentional_homicide_rate}
\item Education (most recent value) \url{https://en.wikipedia.org/wiki/Education_Index}
\item Disposable income \url{https://en.wikipedia.org/wiki/List_of_countries_by_average_wage}
}
}
\note{
Last updated on 2015-12-11
}
\keyword{dataset}

