% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penis.R
\docType{data}
\name{penis}
\alias{penis}
\title{Word Penis Size}
\format{A 146 x 10 \code{data.frame} containing numerical and factor data with the following variables:
\describe{
\item{country}{Country of measurement}
\item{region}{Region of measurement}
\item{length_flaccid}{Flaccid length in cm}
\item{length_erect}{Erect length in cm}
\item{circumf_flaccif}{Flaccid circumference in cm}
\item{circumf_erect}{Erect circumference in cm}
\item{volume_erect}{Erect volum in cm^3}
\item{volume_flaccid}{Flaffid volume in cm^3}
\item{method}{Method of measurement: Self reported & Measured}
\item{n}{Sample size of measurement}
}}
\source{
\url{http://www.everyoneweb.com/worldpenissize}
}
\usage{
penis
}
\description{
Penis data across multiple regions.
}
\keyword{dataset}

